Reframe around positive rights and freedoms
The first draft focused on trying to solve the monetization problem too
directly. As I thought more abuot it, it seemed that the real core
question was: is there an important, positive right to be found in FSF
style statements about software freedom?

My personal answer was yes. Software we can share, study, and modify is
what has enabled the entire growth of the internet. When we decide to
open our software, if we are also intending to have it grow a community,
then at its root we need to recognize that. Take the Rawls point of view
- you are behind the veil of ignorance, you don't know if you're going
to be the original author, a curious kid, or Jeff Bezos. How do we
structure a social contract so that, no matter which role you wind up
playing, you have opportunity, without software freedom at its root?

This re-framing rebuilds the Community Compact to rest its structure
firmly on this fact. The community exists because that right exists,
therefore its sustainability must also drive forward that right. The
Business is now responsible not only to provide for the sustaining
community, it must also provide a built-in mechanism for the community
to decide it is no longer interested in adhering to its terms through
forking. (Hat tip to @laserllama, this was his idea)

In this version, the FREELOADERS file is removed. While it was an
interesting idea, I think it fundamentally works against the idea of
having a healthy community. Either you are a part of the community, and
agree to the terms of the social contract, or you can fork and build a
new community.

It also removes the "nc" builds, for the same reason. If we want a build
that doesn't include the communities sustaining terms, we can create
one, but we can't say it's of the same community.

Clarifies that in return for the right to monetize on behalf of the
community, the Business agrees to only publish software under the terms
of this Community Compact. No open core - fixes #2.

Also clarifies that community is bigger than contributors.

Fixes #8, #7, #3

Signed-off-by: Adam Jacob <adam@stalecoffee.org>