Make sustainability more explicit in core beliefs
After reading through the opening, it struck me as slightly odd that sustainability only came up when tied to the business. Seemed more clear to state it separately (and as a high priority!), so that it can later be used more explicitly as a justification for the business and other rules.

To later support the code of conduct, I used the phrase "economically _and ethically_" sustainable. Open to better suggestions there - could drop the adjectives.

I also added some h3s to help make the structure of this section more clear. This helped me see what looked like some redundancy; so I went ahead and combined the two discussions of forking into one. No real attachment to this part of the change, just seemed cleaner; can back it out if you want. (Could also make a parallel simplification of the first paragraph by moving the discussion of "who the rights apply to" into "who we are".)

Note that the diff is showing a change in "We believe that the software we produce..." paragraph, but I have no idea why - presumably some whitespace somewhere. No actual change that I can see.